fix the initialize bug
